Add a stylish layout for the about-me section on the webpage
On the about-me section of the website, I have styled the header as well as the paragraphs of the inner border. While I added a list of elements for the two paragraphs, I edited these paragraphs various ways such as the lining height for these. I have also edited the margins of these paragraphs within the border so that these can be readable for users that visit the website