Add a a couple of projects to the projects tab of the portfolio website
This tab displays a couple of projects that I have worked on previously. Each project has an external link icon placed at the top left corner of each project where these take visitors to my projects on GitHub. Each includes a short description with a list of technologies that I used to work on for these projects